Extract suppressed region locator into internal file.
Part Ktlint cleanup.
