+++ ktlint-core/src/main/kotlin/com/pinterest/ktlint/core/KtLint.kt
@@ -1,28 +1,21 @@
 package com.pinterest.ktlint.core
 
-import com.pinterest.ktlint.core.ast.prevLeaf
 import com.pinterest.ktlint.core.ast.visit
 import com.pinterest.ktlint.core.internal.EditorConfigLoader
 import com.pinterest.ktlint.core.internal.buildPositionInTextLocator
+import com.pinterest.ktlint.core.internal.buildSuppressedRegionsLocator
 import com.pinterest.ktlint.core.internal.initPsiFileFactory
+import com.pinterest.ktlint.core.internal.noSuppression
 import java.nio.file.FileSystems
 import java.nio.file.Path
 import java.nio.file.Paths
-import java.util.ArrayList
-import java.util.HashSet
 import org.jetbrains.kotlin.com.intellij.lang.ASTNode
 import org.jetbrains.kotlin.com.intellij.openapi.util.Key
-import org.jetbrains.kotlin.com.intellij.psi.PsiComment
 import org.jetbrains.kotlin.com.intellij.psi.PsiElement
 import org.jetbrains.kotlin.com.intellij.psi.PsiErrorElement
 import org.jetbrains.kotlin.com.intellij.psi.PsiFileFactory
-import org.jetbrains.kotlin.com.intellij.psi.PsiWhiteSpace
 import org.jetbrains.kotlin.idea.KotlinLanguage
-import org.jetbrains.kotlin.psi.KtAnnotated
-import org.jetbrains.kotlin.psi.KtAnnotationEntry
 import org.jetbrains.kotlin.psi.KtFile
-import org.jetbrains.kotlin.psi.psiUtil.endOffset
-import org.jetbrains.kotlin.psi.psiUtil.startOffset
 
 object KtLint {
 
@@ -35,8 +28,6 @@ object KtLint {
 
     private val psiFileFactory: PsiFileFactory = initPsiFileFactory()
     private val editorConfigLoader = EditorConfigLoader(FileSystems.getDefault())
-
-    private val nullSuppression = { _: Int, _: String, _: Boolean -> false }
 
     /**
      * @param fileName path of file to lint/format
@@ -92,7 +83,7 @@ object KtLint {
             params.debug
         ) + params.userData
         injectUserData(rootNode, mergedUserData)
-        val isSuppressed = calculateSuppressedRegions(rootNode)
+        val isSuppressed = buildSuppressedRegionsLocator(rootNode)
         val errors = mutableListOf<LintError>()
         visitor(rootNode, params.ruleSets).invoke { node, rule, fqRuleId ->
             // fixme: enforcing suppression based on node.startOffset is wrong
@@ -223,20 +214,6 @@ object KtLint {
         return buildPositionInTextLocator(text)
     }
 
-    private fun calculateSuppressedRegions(rootNode: ASTNode) =
-        SuppressionHint.collect(rootNode).let { listOfHints ->
-            if (listOfHints.isEmpty()) nullSuppression else { offset, ruleId, root ->
-                if (root) {
-                    val h = listOfHints[0]
-                    h.range.endInclusive == 0 && (h.disabledRules.isEmpty() || h.disabledRules.contains(ruleId))
-                } else {
-                    listOfHints.any { (range, disabledRules) ->
-                        (disabledRules.isEmpty() || disabledRules.contains(ruleId)) && range.contains(offset)
-                    }
-                }
-            }
-        }
-
     /**
      * Fix style violations.
      *
@@ -263,7 +240,7 @@ object KtLint {
             params.debug
         ) + params.userData
         injectUserData(rootNode, mergedUserData)
-        var isSuppressed = calculateSuppressedRegions(rootNode)
+        var isSuppressed = buildSuppressedRegionsLocator(rootNode)
         var tripped = false
         var mutated = false
         visitor(rootNode, params.ruleSets, concurrent = false)
@@ -276,8 +253,8 @@ object KtLint {
                             tripped = true
                             if (canBeAutoCorrected) {
                                 mutated = true
-                                if (isSuppressed !== nullSuppression) {
-                                    isSuppressed = calculateSuppressedRegions(rootNode)
+                                if (isSuppressed !== noSuppression) {
+                                    isSuppressed = buildSuppressedRegionsLocator(rootNode)
                                 }
                             }
                         }
@@ -335,121 +312,6 @@ object KtLint {
             eol == "native" -> System.lineSeparator()
             eol == "crlf" || eol != "lf" && fileContent.lastIndexOf('\r') != -1 -> "\r\n"
             else -> "\n"
-        }
-    }
-
-    /**
-     * @param range zero-based range of lines where lint errors should be suppressed
-     * @param disabledRules empty set means "all"
-     */
-    private data class SuppressionHint(val range: IntRange, val disabledRules: Set<String> = emptySet()) {
-
-        companion object {
-
-            fun collect(rootNode: ASTNode): List<SuppressionHint> {
-                val result = ArrayList<SuppressionHint>()
-                val open = ArrayList<SuppressionHint>()
-                rootNode.visit { node ->
-                    if (node is PsiComment) {
-                        val text = node.getText()
-                        if (text.startsWith("//")) {
-                            val commentText = text.removePrefix("//").trim()
-                            parseHintArgs(commentText, "ktlint-disable")?.let { args ->
-                                val lineStart = (
-                                    node.prevLeaf { it is PsiWhiteSpace && it.textContains('\n') } as
-                                        PsiWhiteSpace?
-                                    )?.let { it.node.startOffset + it.text.lastIndexOf('\n') + 1 } ?: 0
-                                result.add(SuppressionHint(IntRange(lineStart, node.startOffset), HashSet(args)))
-                            }
-                        } else {
-                            val commentText = text.removePrefix("/*").removeSuffix("*/").trim()
-                            parseHintArgs(commentText, "ktlint-disable")?.apply {
-                                open.add(SuppressionHint(IntRange(node.startOffset, node.startOffset), HashSet(this)))
-                            }
-                                ?: parseHintArgs(commentText, "ktlint-enable")?.apply {
-                                    // match open hint
-                                    val disabledRules = HashSet(this)
-                                    val openHintIndex = open.indexOfLast { it.disabledRules == disabledRules }
-                                    if (openHintIndex != -1) {
-                                        val openingHint = open.removeAt(openHintIndex)
-                                        result.add(
-                                            SuppressionHint(
-                                                IntRange(openingHint.range.start, node.startOffset),
-                                                disabledRules
-                                            )
-                                        )
-                                    }
-                                }
-                        }
-                    }
-                    // Extract all Suppress annotations and create SuppressionHints
-                    val psi = node.psi
-                    if (psi is KtAnnotated) {
-                        createSuppressionHintFromAnnotations(psi, suppressAnnotations, suppressAnnotationRuleMap)
-                            ?.let {
-                                result.add(it)
-                            }
-                    }
-                }
-                result.addAll(
-                    open.map {
-                        SuppressionHint(IntRange(it.range.first, rootNode.textLength), it.disabledRules)
-                    }
-                )
-                return result
-            }
-
-            private fun parseHintArgs(commentText: String, key: String): List<String>? {
-                if (commentText.startsWith(key)) {
-                    val parsedComment = splitCommentBySpace(commentText)
-                    // assert exact match
-                    if (parsedComment[0] == key) {
-                        return parsedComment.tail()
-                    }
-                }
-                return null
-            }
-
-            private fun splitCommentBySpace(comment: String) =
-                comment.replace(Regex("\\s"), " ").replace(" {2,}", " ").split(" ")
-
-            private fun <T> List<T>.tail() = this.subList(1, this.size)
-
-            private val suppressAnnotationRuleMap = mapOf(
-                "RemoveCurlyBracesFromTemplate" to "string-template"
-            )
-
-            private val suppressAnnotations = setOf("Suppress", "SuppressWarnings")
-
-            /**
-             * Creates [SuppressionHint] from annotations of given [psi]
-             * Returns null if no [targetAnnotations] present or no mapping exists
-             * between annotations' values and ktlint rules
-             */
-            private fun createSuppressionHintFromAnnotations(
-                psi: KtAnnotated,
-                targetAnnotations: Collection<String>,
-                annotationValueToRuleMapping: Map<String, String>
-            ): SuppressionHint? {
-
-                return psi.annotationEntries
-                    .filter {
-                        it.calleeExpression?.constructorReferenceExpression
-                            ?.getReferencedName() in targetAnnotations
-                    }.flatMap(KtAnnotationEntry::getValueArguments)
-                    .mapNotNull { it.getArgumentExpression()?.text?.removeSurrounding("\"") }
-                    .mapNotNull(annotationValueToRuleMapping::get)
-                    .let { suppressedRules ->
-                        if (suppressedRules.isNotEmpty()) {
-                            SuppressionHint(
-                                IntRange(psi.startOffset, psi.endOffset),
-                                suppressedRules.toSet()
-                            )
-                        } else {
-                            null
-                        }
-                    }
-            }
         }
     }
